Cleanup main.py by adding more utils
